Move `filter_out_duplicates` function
This PR moves the `filter_out_duplicates()` frunction from the `gilda.generate_terms` module to `gilda.term` module.

This makes it possible for `biomappings` to import this function without installing `indra`.

Since `filter_out_duplicates` is still imported by `gilda.generated_terms`, it is still available in the namespace and therefore no downstream code needs to be modified.
